+++ errorHandlers/index.js
@@ -1,0 +1,2 @@
+export { errorOccered } from "./errorOccered";
+export { notFound } from "./routeNotFound";
