+++ errorHandlers/routeNotFound.js
@@ -1,0 +1,8 @@
+export const notFound = (req, res) => {
+  res
+    .status(404)
+    .json({
+      success: false,
+      message: "route not found on server, please check",
+    });
+};
