+++ routes/watch-later.js
@@ -1,0 +1,20 @@
+export const postWatchLater = async (req, res) => {
+  let { user } = req;
+  const watchlater = req.body;
+  user.watch_later.push(watchlater);
+  await user.save();
+  res.json({ success: true });
+};
+
+export const deleteWatchLater = async (req, res) => {
+  let { user } = req;
+  const watch_later_id = req.body.id;
+  const updatedLater = user.watch_later.filter(
+    (video) => video.id !== watch_later_id
+  );
+  user = extend(user, { watch_later: updatedLater });
+
+  const NewUser = User(user);
+  const savedUser = await NewUser.save();
+  res.json({ success: true });
+};
